requirements.txt
flask==3.0.0
pandas==2.1.4
requests==2.31.0
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
werkzeug==3.0.1